Refactor: move buildConfirmMessage to from-sdk Fixes: AFORM-3920
